Extract basic boilerplate template
This extracts basic, minimal functionality and setup from some of my existing
applications into this generator.

The implementation is inspired by [suspenders] and [schienenzeppelin], with
functionality captured in custom `AppGenerator` and `AppBuilder` classes.

Since the initial functionality is limited and the generator is not
configurable, there is not yet a need for any more granular abstractions.

[suspenders]: https://github.com/thoughtbot/suspenders
[schienenzeppelin]: https://github.com/hschne/schienenzeppelin